fix: move more types to correct place
Closes https://github.com/kad-products/rezept-core/issues/140
