+++ src/types/recipe-ingredients.ts
@@ -5,3 +5,8 @@ export type RecipeIngredientFormSave = Omit<
 	typeof recipeIngredients.$inferInsert,
 	'recipeSectionId' | 'createdAt' | 'createdBy' | 'updatedAt' | 'updatedBy' | 'deletedAt' | 'deletedBy'
 >;
+
+export type IncomingIngredientsData = {
+	sectionId: string;
+	ingredients: RecipeIngredientFormSave[] | undefined;
+};
