+++ src/types/recipe-instructions.ts
@@ -5,3 +5,8 @@ export type RecipeInstructionFormSave = Omit<
 	typeof recipeInstructions.$inferInsert,
 	'recipeSectionId' | 'createdAt' | 'createdBy' | 'updatedAt' | 'updatedBy' | 'deletedAt' | 'deletedBy'
 >;
+
+export type IncomingInstructionsData = {
+	sectionId: string;
+	instructions: RecipeInstructionFormSave[] | undefined;
+};
