+++ src/types/recipe-scrapes.ts
@@ -1,5 +1,21 @@
 import type { recipeScrapes } from '@/models';
 
+export type ParsedRecipeScrapeSection = {
+	order: number;
+	ingredients: { raw: string; order: number }[];
+	instructions: { stepNumber: number; instruction: string }[];
+};
+
+export type ParsedRecipeScrape = {
+	title: string;
+	description?: string;
+	source?: string;
+	servings?: number;
+	prepTime?: number;
+	cookTime?: number;
+	sections: ParsedRecipeScrapeSection[];
+};
+
 // should match the model's enum
 export type RecipeScrapeStatus =
 	| 'SCRAPED'
